+++ nix/modules/project/settings/extraBuildDepends.nix
@@ -1,0 +1,19 @@
+{ config, pkgs, lib, mkCabalSettingOptions, ... }:
+
+let
+  inherit (lib)
+    types;
+  inherit (types)
+    listOf;
+in
+{
+  options = mkCabalSettingOptions {
+    inherit config;
+    name = "extraBuildDepends";
+    type = listOf types.package;
+    description = ''
+      Extra build dependencies for the package.
+    '';
+    impl = pkgs.haskell.lib.compose.addBuildDepends;
+  };
+}